moving screen files to appropriate folders
